[package] update monit from 5.0.3 to 5.1.1
I wanted to configure monit with a rule that checks a network
connection, restarts the connection if it fails, and eventually
reboots the entire router if the connection restarts don't work.

I wrote this in my rules:

  start "/sbin/ifup inet"
  stop "/sbin/ifdown inet"
  if failed icmp type echo count 3 timeout 10 seconds then restart
  if 3 restarts within 5 cycles then exec "/sbin/reboot"

When I tested these rules I found that monit 5.0.3 does not
support the phrase:

  ... then exec "/sbin/reboot"

But this phrase is supported by monit 5.1.1.

I patched the Makefile to upgrade to 5.1.1, rebuilt and reflashed,
and now my monit rule runs as intended.  So I have tested this on a
grand total of one machine (Linksys WAG354G, TI AR7).  Is this
enough to justify upgrading monit for everyone?

My changes to package/feeds/packages/monit/Makefile are below.
They were generated with svn diff.

Signed-off-by: Jon Hirst <openwrt@jonmail.co.uk>

git-svn-id: svn+ssh://svn.openwrt.org/openwrt/packages@23624 3c298f89-4303-0410-b956-a3cf2f4a3e73