avahi: cleanup / enhance
This patch updates the avahi Makefile to correct a number of errors. In
particular it handles the choice of D-BUS- or non-D-BUS-enabled packages
more consistently.

Packages that don't require D-BUS support are compiled with D-BUS
support disabled by default, keeping the build as small as possible.
Support for D-BUS can be selected if desired.

Selecting packages that require D-BUS support -- avahi-utils or
libavahi-client -- will automatically cause all packages to be built
with D-BUS support and will add the D-BUS package to the build.

Note that after this update, the D-BUS package will be needed as part of
the compilation environment, so it will be downloaded and compiled, but
it will not be added to the build unless the user has enabled D-BUS
support or has selected avahi-utils or the avahi-client library, both of
which need it.

Tested on trunk r29283.

Signed-off-by Mike Brady <mikebrady@eircom.net>

git-svn-id: svn+ssh://svn.openwrt.org/openwrt/packages@29359 3c298f89-4303-0410-b956-a3cf2f4a3e73