v1 : do not install omshell static librairies in dhcp package
There is no need to install any static library on the target, unless one installs a linker (and a compiler) as well, and this is certainly not a common case.
The dhcp-server package installs two of them in /usr/local/lib - along with omshell. The erroneous lines were added as part of r10146 when the Makefile
was modified to install omshell, but the static libraries are never needed to run omshell on the target.

Signed-off-by: Emmanuel Deloget <logout@free.fr>

git-svn-id: svn+ssh://svn.openwrt.org/openwrt/packages@30707 3c298f89-4303-0410-b956-a3cf2f4a3e73