avahi: do not announce a specific front page
Announcing /index.html as frontpage is right when luci or x-wrt is installed
as web gui since both applications come with an /index.html which redirect afterwards.

However, when using an other front page, e.g. an /index.php then the announcement
is wrong and results in a 404 Not Found.
So I wonder/suggest whether the default avahi service could not announce
such a specific frontpage but only the generic "top level" directory.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>

git-svn-id: svn+ssh://svn.openwrt.org/openwrt/packages@31375 3c298f89-4303-0410-b956-a3cf2f4a3e73