radsecproxy: update to 1.6.2
Since my previous post a similar issue was discovered for DTLS and 1.6.2
was released. CVE numbers were also assigned for these two
vulnerabilities:

CVE-2012-4523 refers to the TLS issue, fixed in radsecproxy 1.6.1
CVE-2012-4566 refers to the DTLS issue, fixed in radsecproxy 1.6.2

So it would be good to bump to 1.6.2, applying this patch on top of
r34484.

Signed-off-by: Zenon Mousmoulas <zmousm@noc.grnet.gr>
Signed-off-by: Florian Fainelli <florian@openwrt.org>

git-svn-id: svn+ssh://svn.openwrt.org/openwrt/packages@34535 3c298f89-4303-0410-b956-a3cf2f4a3e73