collectd: remove bogus maximum of 3600 seconds for 'timeleft'
The apcups and snmp plugins were using this to store time in minutes, but
nut uses seconds. Upstream wants to make them all use seconds (unlike my
first patch to make nut use minutes to be consistent with the others), so
all we need to do is fix the maximum. If we were *using* the apcups and
snmp plugins, those would probably want fixing to multiply by 60.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>
Signed-off-by: Gabor Juhos <juhosg@openwrt.org>

git-svn-id: svn+ssh://svn.openwrt.org/openwrt/packages@34831 3c298f89-4303-0410-b956-a3cf2f4a3e73