libdlna: version bump to 0.2.4
Update libdlna to 0.2.4.

Drop 100-configure_fix.patch. This was a mix of configure fixes and changes
to build with newer ffmpeg versions. The ffmpeg changes are integrated.
The configure changes are rolled into a new patch, 010-configure-crosscompile.patch,
so it will hopefully stay single purpose changes.

Rebase 300-ffmpeg_compat.patch.

Note: This is compile tested only.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>
Signed-off-by: Gabor Juhos <juhosg@openwrt.org>

git-svn-id: svn+ssh://svn.openwrt.org/openwrt/packages@35796 3c298f89-4303-0410-b956-a3cf2f4a3e73