[packages] tinc: Version bump 1.0.20
With pleasure we announce the release of tinc version 1.0.20. Here is a
summary of the changes:

 * Use /dev/tap0 by default on FreeBSD and NetBSD when using switch mode.

 * Minor improvements and clarifications in the documentation.

 * Allow tinc to be cross-compiled with Android's NDK.

 * The discovered PMTU is now also applied to VLAN tagged traffic.

 * The LocalDiscovery option now makes use of all addresses tinc is bound to.

 * Fixed support for tunemu on iOS devices.

 * The PriorityInheritance option now also works with switch mode.

 * Fixed tinc crashing when using a SOCKS5 proxy.

Thanks to Mesar Hameed, Vilbrekin and Martin Schürrer for their contributions
to this version of tinc.

This version of tinc is compatible with 1.0pre8, 1.0 and later, but not
with earlier version of tinc.

Note that this will be one of the last releases of the 1.0 branch of tinc. Most
of the development is now happening on the 1.1 branch, with tinc 1.1pre6 being
the latest pre-release version, which you are encouraged to try out. The 1.1
branch is backwards compatible with tinc 1.0.

git-svn-id: svn+ssh://svn.openwrt.org/openwrt/packages@36041 3c298f89-4303-0410-b956-a3cf2f4a3e73